Split selected attention Triton implementation
