--- attn_gym/sparse/selected_attention/impl/triton.py
@@ -1,1433 +1,0 @@
-"""Triton backend for selected attention."""
-
-import math
-
-import torch
-import triton
-import triton.language as tl
-from triton.tools.tensor_descriptor import TensorDescriptor
-
-from attn_gym._backends.triton.utils import can_use_tma, ptr_offset
-
-
-@triton.jit
-def _load_bhsd(
-    tensor_ptr,
-    strides: tl.constexpr,
-    batch,
-    head,
-    positions,
-    offsets_d,
-    mask,
-):
-    """Load a tile from a tensor in batch-head-sequence-dimension order."""
-    return tl.load(
-        tensor_ptr
-        + ptr_offset(
-            (batch, head, positions[:, None], offsets_d[None, :]),
-            strides,
-        ),
-        mask=mask,
-        other=0.0,
-    )
-
-
-@triton.jit
-def _load_bs(
-    tensor_ptr,
-    strides: tl.constexpr,
-    batch,
-    positions,
-    mask,
-    other: tl.constexpr,
-):
-    """Load positions from a tensor in batch-sequence order."""
-    return tl.load(
-        tensor_ptr + ptr_offset((batch, positions), strides),
-        mask=mask,
-        other=other,
-    )
-
-
-@triton.jit
-def _store_bhsd(
-    tensor_ptr,
-    value,
-    strides: tl.constexpr,
-    batch,
-    head,
-    positions,
-    offsets_d,
-    mask,
-):
-    """Store a tile to a tensor in batch-head-sequence-dimension order."""
-    tl.store(
-        tensor_ptr
-        + ptr_offset(
-            (batch, head, positions[:, None], offsets_d[None, :]),
-            strides,
-        ),
-        value,
-        mask=mask,
-    )
-
-
-@triton.jit
-def _merge_attention_tile(accumulator, running_max, running_sum, logits, values):
-    """Merge a tile while keeping online-softmax state and accumulation in FP32."""
-    tile_max = tl.max(logits, axis=1)
-    new_max = tl.maximum(running_max, tile_max)
-    alpha = tl.exp(running_max - new_max)
-    probabilities = tl.exp(logits - new_max[:, None])
-    accumulator *= alpha[:, None]
-    accumulator += tl.dot(probabilities.to(values.dtype), values, input_precision="tf32x3")
-    running_sum = running_sum * alpha + tl.sum(probabilities, axis=1)
-    return accumulator, new_max, running_sum
-
-
-@triton.jit
-def _selected_attention_fwd(
-    query_ptr,
-    sparse_kv_ptr,
-    local_kv_ptr,
-    kv_indices_ptr,
-    doc_ids_ptr,
-    attention_sink_ptr,
-    output_ptr,
-    lse_ptr,
-    QUERY_STRIDES: tl.constexpr,
-    SPARSE_KV_STRIDES: tl.constexpr,
-    LOCAL_KV_STRIDES: tl.constexpr,
-    KV_INDICES_STRIDES: tl.constexpr,
-    DOC_IDS_STRIDES: tl.constexpr,
-    LSE_STRIDES: tl.constexpr,
-    H: tl.constexpr,
-    S: tl.constexpr,
-    D: tl.constexpr,
-    SPARSE_SEQ_LEN: tl.constexpr,
-    TOPK: tl.constexpr,
-    WINDOW: tl.constexpr,
-    SCALE: tl.constexpr,
-    HAS_DOC_IDS: tl.constexpr,
-    NUM_LOCAL_TILES: tl.constexpr,
-    BLOCK_M: tl.constexpr,
-    BLOCK_N: tl.constexpr,
-    BLOCK_D: tl.constexpr,
-):
-    """Apply online softmax over selected sparse entries and the local window."""
-    query_block = tl.program_id(0)
-    batch_head = tl.program_id(1)
-    head = batch_head % H
-    batch = batch_head // H
-
-    offsets_m = query_block * BLOCK_M + tl.arange(0, BLOCK_M)
-    offsets_d = tl.arange(0, BLOCK_D)
-    query_mask = offsets_m < S
-    dimension_mask = offsets_d < D
-
-    query = _load_bhsd(
-        query_ptr,
-        QUERY_STRIDES,
-        batch,
-        head,
-        offsets_m,
-        offsets_d,
-        query_mask[:, None] & dimension_mask[None, :],
-    )
-
-    if HAS_DOC_IDS:
-        query_doc_ids = _load_bs(doc_ids_ptr, DOC_IDS_STRIDES, batch, offsets_m, query_mask, -1)
-
-    sink = tl.load(attention_sink_ptr + head).to(tl.float32)
-    running_max = tl.full((BLOCK_M,), sink, tl.float32)
-    running_sum = tl.full((BLOCK_M,), 1.0, tl.float32)
-    accumulator = tl.zeros((BLOCK_M, BLOCK_D), tl.float32)
-
-    # Each query has its own set of selected sparse entries.
-    for selected_slot in tl.static_range(0, TOPK):
-        selected_idx = tl.load(
-            kv_indices_ptr
-            + batch * KV_INDICES_STRIDES[0]
-            + offsets_m * KV_INDICES_STRIDES[1]
-            + selected_slot * KV_INDICES_STRIDES[2],
-            mask=query_mask,
-            other=0,
-        )
-        valid = query_mask & (selected_idx >= 0) & (selected_idx < SPARSE_SEQ_LEN)
-        sparse_value = _load_bhsd(
-            sparse_kv_ptr,
-            SPARSE_KV_STRIDES,
-            batch,
-            head,
-            selected_idx,
-            offsets_d,
-            valid[:, None] & dimension_mask[None, :],
-        )
-        logit = tl.sum(query * sparse_value, axis=1) * SCALE
-        logit = tl.where(valid, logit, -float("inf"))
-        new_max = tl.maximum(running_max, logit)
-        alpha = tl.exp(running_max - new_max)
-        probability = tl.exp(logit - new_max)
-        accumulator = accumulator * alpha[:, None] + probability[:, None] * sparse_value
-        running_sum = running_sum * alpha + probability
-        running_max = new_max
-
-    # The local window is processed in tensor-core-friendly tiles.
-    first_local_position = query_block * BLOCK_M - WINDOW + 1
-    offsets_n_base = tl.arange(0, BLOCK_N)
-    for local_tile in tl.static_range(0, NUM_LOCAL_TILES):
-        local_start = first_local_position + local_tile * BLOCK_N
-        offsets_n = local_start + offsets_n_base
-        local_mask = (offsets_n >= 0) & (offsets_n < S)
-        local_values = _load_bhsd(
-            local_kv_ptr,
-            LOCAL_KV_STRIDES,
-            batch,
-            head,
-            offsets_n,
-            offsets_d,
-            local_mask[:, None] & dimension_mask[None, :],
-        )
-        logits = tl.dot(query, tl.trans(local_values), input_precision="tf32x3") * SCALE
-        causal_window_mask = (
-            query_mask[:, None]
-            & local_mask[None, :]
-            & (offsets_n[None, :] <= offsets_m[:, None])
-            & (offsets_n[None, :] >= offsets_m[:, None] - WINDOW + 1)
-        )
-        if HAS_DOC_IDS:
-            key_doc_ids = _load_bs(doc_ids_ptr, DOC_IDS_STRIDES, batch, offsets_n, local_mask, -2)
-            doc_mask = query_doc_ids[:, None] == key_doc_ids[None, :]
-            causal_window_mask = causal_window_mask & doc_mask
-
-        logits = tl.where(causal_window_mask, logits, -float("inf"))
-        accumulator, running_max, running_sum = _merge_attention_tile(
-            accumulator, running_max, running_sum, logits, local_values
-        )
-
-    output = accumulator / running_sum[:, None]
-    _store_bhsd(
-        output_ptr,
-        output,
-        QUERY_STRIDES,
-        batch,
-        head,
-        offsets_m,
-        offsets_d,
-        query_mask[:, None] & dimension_mask[None, :],
-    )
-    tl.store(
-        lse_ptr + batch * LSE_STRIDES[0] + head * LSE_STRIDES[1] + offsets_m * LSE_STRIDES[2],
-        running_max + tl.log(running_sum),
-        mask=query_mask,
-    )
-
-
-@triton.autotune(
-    configs=[
-        triton.Config({"BLOCK_N": block_n}, num_warps=num_warps, num_stages=1)
-        for block_n in (64, 128, 256)
-        for num_warps in (4, 8)
-    ],
-    key=["B", "H", "S", "D", "SPARSE_SEQ_LEN", "TOPK", "WINDOW", "HAS_DOC_IDS"],
-    cache_results=True,
-)
-@triton.jit
-def _selected_attention_fwd_shared(
-    query_ptr,
-    sparse_kv_ptr,
-    local_kv_ptr,
-    kv_indices_ptr,
-    doc_ids_ptr,
-    attention_sink_ptr,
-    output_ptr,
-    lse_ptr,
-    QUERY_STRIDES: tl.constexpr,
-    SPARSE_KV_STRIDES: tl.constexpr,
-    LOCAL_KV_STRIDES: tl.constexpr,
-    KV_INDICES_STRIDES: tl.constexpr,
-    DOC_IDS_STRIDES: tl.constexpr,
-    LSE_STRIDES: tl.constexpr,
-    B: tl.constexpr,
-    H: tl.constexpr,
-    S: tl.constexpr,
-    D: tl.constexpr,
-    SPARSE_SEQ_LEN: tl.constexpr,
-    TOPK: tl.constexpr,
-    WINDOW: tl.constexpr,
-    SCALE: tl.constexpr,
-    HAS_DOC_IDS: tl.constexpr,
-    BLOCK_H: tl.constexpr,
-    BLOCK_K: tl.constexpr,
-    BLOCK_N: tl.constexpr,
-    BLOCK_D: tl.constexpr,
-):
-    """Process one sequence position across heads to reuse shared KV tiles."""
-    sequence = tl.program_id(0)
-    batch = tl.program_id(1)
-    head_block = tl.program_id(2)
-
-    offsets_h = head_block * BLOCK_H + tl.arange(0, BLOCK_H)
-    offsets_d = tl.arange(0, BLOCK_D)
-    head_mask = offsets_h < H
-    dimension_mask = offsets_d < D
-
-    query = tl.load(
-        query_ptr
-        + ptr_offset(
-            (batch, offsets_h[:, None], sequence, offsets_d[None, :]),
-            QUERY_STRIDES,
-        ),
-        mask=head_mask[:, None] & dimension_mask[None, :],
-        other=0.0,
-    )
-    sink = tl.load(attention_sink_ptr + offsets_h, mask=head_mask, other=0.0).to(tl.float32)
-    running_max = sink
-    running_sum = tl.full((BLOCK_H,), 1.0, tl.float32)
-    accumulator = tl.zeros((BLOCK_H, BLOCK_D), tl.float32)
-
-    if TOPK:
-        offsets_k = tl.arange(0, BLOCK_K)
-        selected_idx = tl.load(
-            kv_indices_ptr + ptr_offset((batch, sequence, offsets_k), KV_INDICES_STRIDES),
-            mask=offsets_k < TOPK,
-            other=-1,
-        )
-        selected_valid = (offsets_k < TOPK) & (selected_idx >= 0) & (selected_idx < SPARSE_SEQ_LEN)
-        selected_idx = tl.where(selected_valid, selected_idx, 0)
-        sparse_values = tl.load(
-            sparse_kv_ptr
-            + ptr_offset(
-                (batch, 0, selected_idx[:, None], offsets_d[None, :]),
-                SPARSE_KV_STRIDES,
-            ),
-            mask=selected_valid[:, None] & dimension_mask[None, :],
-            other=0.0,
-        )
-        logits = tl.dot(query, tl.trans(sparse_values), input_precision="tf32x3") * SCALE
-        logits = tl.where(head_mask[:, None] & selected_valid[None, :], logits, -float("inf"))
-        accumulator, running_max, running_sum = _merge_attention_tile(
-            accumulator, running_max, running_sum, logits, sparse_values
-        )
-
-    if HAS_DOC_IDS:
-        query_doc_id = tl.load(doc_ids_ptr + ptr_offset((batch, sequence), DOC_IDS_STRIDES))
-
-    offsets_n_base = tl.arange(0, BLOCK_N)
-    first_local_position = sequence - WINDOW + 1
-    num_local_tiles: tl.constexpr = tl.cdiv(WINDOW, BLOCK_N)
-    for local_tile in tl.static_range(0, num_local_tiles):
-        offsets_n = first_local_position + local_tile * BLOCK_N + offsets_n_base
-        local_valid = (offsets_n >= 0) & (offsets_n <= sequence) & (offsets_n < S)
-        local_values = tl.load(
-            local_kv_ptr
-            + ptr_offset(
-                (batch, 0, offsets_n[:, None], offsets_d[None, :]),
-                LOCAL_KV_STRIDES,
-            ),
-            mask=local_valid[:, None] & dimension_mask[None, :],
-            other=0.0,
-        )
-        if HAS_DOC_IDS:
-            key_doc_ids = tl.load(
-                doc_ids_ptr + ptr_offset((batch, offsets_n), DOC_IDS_STRIDES),
-                mask=local_valid,
-                other=-1,
-            )
-            local_valid &= key_doc_ids == query_doc_id
-
-        logits = tl.dot(query, tl.trans(local_values), input_precision="tf32x3") * SCALE
-        logits = tl.where(head_mask[:, None] & local_valid[None, :], logits, -float("inf"))
-        accumulator, running_max, running_sum = _merge_attention_tile(
-            accumulator, running_max, running_sum, logits, local_values
-        )
-
-    tl.store(
-        output_ptr
-        + ptr_offset(
-            (batch, offsets_h[:, None], sequence, offsets_d[None, :]),
-            QUERY_STRIDES,
-        ),
-        accumulator / running_sum[:, None],
-        mask=head_mask[:, None] & dimension_mask[None, :],
-    )
-    tl.store(
-        lse_ptr + ptr_offset((batch, offsets_h, sequence), LSE_STRIDES),
-        running_max + tl.log(running_sum),
-        mask=head_mask,
-    )
-
-
-@triton.autotune(
-    configs=[
-        triton.Config({}, num_warps=num_warps, num_stages=num_stages)
-        for num_warps in (4, 8)
-        for num_stages in (1, 3)
-    ],
-    key=["H", "S", "D", "SPARSE_SEQ_LEN", "TOPK", "WINDOW", "HAS_DOC_IDS"],
-    cache_results=True,
-)
-@triton.jit
-def _selected_attention_fwd_tma(
-    query_desc,
-    sparse_kv_ptr,
-    local_desc,
-    kv_indices_ptr,
-    doc_ids_ptr,
-    attention_sink_ptr,
-    output_desc,
-    lse_ptr,
-    SPARSE_KV_STRIDES: tl.constexpr,
-    KV_INDICES_STRIDES: tl.constexpr,
-    DOC_IDS_STRIDES: tl.constexpr,
-    LSE_STRIDES: tl.constexpr,
-    H: tl.constexpr,
-    S: tl.constexpr,
-    D: tl.constexpr,
-    SPARSE_SEQ_LEN: tl.constexpr,
-    TOPK: tl.constexpr,
-    WINDOW: tl.constexpr,
-    SCALE: tl.constexpr,
-    HAS_DOC_IDS: tl.constexpr,
-    NUM_LOCAL_TILES: tl.constexpr,
-    BLOCK_M: tl.constexpr,
-    BLOCK_N: tl.constexpr,
-    BLOCK_D: tl.constexpr,
-):
-    """TMA forward using host-created descriptors for dense tiles."""
-    query_block = tl.program_id(0)
-    batch_head = tl.program_id(1)
-    head = batch_head % H
-    batch = batch_head // H
-
-    offsets_m = query_block * BLOCK_M + tl.arange(0, BLOCK_M)
-    offsets_d = tl.arange(0, BLOCK_D)
-    query_mask = offsets_m < S
-    dimension_mask = offsets_d < D
-    query = tl.reshape(
-        query_desc.load([batch, head, query_block * BLOCK_M, 0]),
-        (BLOCK_M, BLOCK_D),
-    )
-
-    if HAS_DOC_IDS:
-        query_doc_ids = _load_bs(doc_ids_ptr, DOC_IDS_STRIDES, batch, offsets_m, query_mask, -1)
-
-    sink = tl.load(attention_sink_ptr + head).to(tl.float32)
-    running_max = tl.full((BLOCK_M,), sink, tl.float32)
-    running_sum = tl.full((BLOCK_M,), 1.0, tl.float32)
-    accumulator = tl.zeros((BLOCK_M, BLOCK_D), tl.float32)
-
-    for selected_slot in tl.static_range(0, TOPK):
-        selected_idx = tl.load(
-            kv_indices_ptr
-            + batch * KV_INDICES_STRIDES[0]
-            + offsets_m * KV_INDICES_STRIDES[1]
-            + selected_slot * KV_INDICES_STRIDES[2],
-            mask=query_mask,
-            other=0,
-        )
-        valid = query_mask & (selected_idx >= 0) & (selected_idx < SPARSE_SEQ_LEN)
-        sparse_value = _load_bhsd(
-            sparse_kv_ptr,
-            SPARSE_KV_STRIDES,
-            batch,
-            head,
-            selected_idx,
-            offsets_d,
-            valid[:, None] & dimension_mask[None, :],
-        )
-        logit = tl.sum(query * sparse_value, axis=1) * SCALE
-        logit = tl.where(valid, logit, -float("inf"))
-        new_max = tl.maximum(running_max, logit)
-        alpha = tl.exp(running_max - new_max)
-        probability = tl.exp(logit - new_max)
-        accumulator = accumulator * alpha[:, None] + probability[:, None] * sparse_value
-        running_sum = running_sum * alpha + probability
-        running_max = new_max
-
-    first_local_position = query_block * BLOCK_M - WINDOW + 1
-    offsets_n_base = tl.arange(0, BLOCK_N)
-    for local_tile in tl.static_range(0, NUM_LOCAL_TILES):
-        local_start = first_local_position + local_tile * BLOCK_N
-        offsets_n = local_start + offsets_n_base
-        local_mask = (offsets_n >= 0) & (offsets_n < S)
-        local_values = tl.reshape(
-            local_desc.load([batch, head, local_start, 0]),
-            (BLOCK_N, BLOCK_D),
-        )
-        logits = tl.dot(query, tl.trans(local_values), input_precision="tf32x3") * SCALE
-        valid = (
-            query_mask[:, None]
-            & local_mask[None, :]
-            & (offsets_n[None, :] <= offsets_m[:, None])
-            & (offsets_n[None, :] >= offsets_m[:, None] - WINDOW + 1)
-        )
-        if HAS_DOC_IDS:
-            key_doc_ids = _load_bs(doc_ids_ptr, DOC_IDS_STRIDES, batch, offsets_n, local_mask, -2)
-            valid &= query_doc_ids[:, None] == key_doc_ids[None, :]
-
-        logits = tl.where(valid, logits, -float("inf"))
-        accumulator, running_max, running_sum = _merge_attention_tile(
-            accumulator, running_max, running_sum, logits, local_values
-        )
-
-    output_desc.store(
-        [batch, head, query_block * BLOCK_M, 0],
-        tl.reshape(accumulator / running_sum[:, None], (1, 1, BLOCK_M, BLOCK_D)),
-    )
-    tl.store(
-        lse_ptr + batch * LSE_STRIDES[0] + head * LSE_STRIDES[1] + offsets_m * LSE_STRIDES[2],
-        running_max + tl.log(running_sum),
-        mask=query_mask,
-    )
-
-
-@triton.jit
-def _selected_attention_bwd_dq(
-    query_ptr,
-    sparse_kv_ptr,
-    local_kv_ptr,
-    kv_indices_ptr,
-    doc_ids_ptr,
-    output_ptr,
-    grad_output_ptr,
-    lse_ptr,
-    attention_sink_ptr,
-    grad_query_ptr,
-    grad_sink_ptr,
-    QUERY_STRIDES: tl.constexpr,
-    SPARSE_KV_STRIDES: tl.constexpr,
-    LOCAL_KV_STRIDES: tl.constexpr,
-    KV_INDICES_STRIDES: tl.constexpr,
-    DOC_IDS_STRIDES: tl.constexpr,
-    LSE_STRIDES: tl.constexpr,
-    H: tl.constexpr,
-    S: tl.constexpr,
-    D: tl.constexpr,
-    SPARSE_SEQ_LEN: tl.constexpr,
-    TOPK: tl.constexpr,
-    WINDOW: tl.constexpr,
-    SCALE: tl.constexpr,
-    HAS_DOC_IDS: tl.constexpr,
-    NUM_LOCAL_TILES: tl.constexpr,
-    BLOCK_M: tl.constexpr,
-    BLOCK_N: tl.constexpr,
-    BLOCK_D: tl.constexpr,
-):
-    """Accumulate query and sink gradients over both attention branches."""
-    query_block = tl.program_id(0)
-    batch_head = tl.program_id(1)
-    head = batch_head % H
-    batch = batch_head // H
-
-    offsets_m = query_block * BLOCK_M + tl.arange(0, BLOCK_M)
-    offsets_d = tl.arange(0, BLOCK_D)
-    query_mask = offsets_m < S
-    dimension_mask = offsets_d < D
-    matrix_mask = query_mask[:, None] & dimension_mask[None, :]
-
-    query = _load_bhsd(query_ptr, QUERY_STRIDES, batch, head, offsets_m, offsets_d, matrix_mask)
-    output = _load_bhsd(output_ptr, QUERY_STRIDES, batch, head, offsets_m, offsets_d, matrix_mask)
-    grad_output = _load_bhsd(
-        grad_output_ptr, QUERY_STRIDES, batch, head, offsets_m, offsets_d, matrix_mask
-    )
-    lse = tl.load(
-        lse_ptr + batch * LSE_STRIDES[0] + head * LSE_STRIDES[1] + offsets_m * LSE_STRIDES[2],
-        mask=query_mask,
-        other=0.0,
-    )
-    delta = tl.sum(grad_output * output, axis=1)
-    grad_query = tl.zeros((BLOCK_M, BLOCK_D), tl.float32)
-
-    if HAS_DOC_IDS:
-        query_doc_ids = _load_bs(doc_ids_ptr, DOC_IDS_STRIDES, batch, offsets_m, query_mask, -1)
-
-    for selected_slot in tl.range(0, TOPK):
-        selected_idx = tl.load(
-            kv_indices_ptr
-            + batch * KV_INDICES_STRIDES[0]
-            + offsets_m * KV_INDICES_STRIDES[1]
-            + selected_slot * KV_INDICES_STRIDES[2],
-            mask=query_mask,
-            other=0,
-        )
-        valid = query_mask & (selected_idx >= 0) & (selected_idx < SPARSE_SEQ_LEN)
-        sparse_value = _load_bhsd(
-            sparse_kv_ptr,
-            SPARSE_KV_STRIDES,
-            batch,
-            head,
-            selected_idx,
-            offsets_d,
-            valid[:, None] & dimension_mask[None, :],
-        )
-        scores = tl.sum(query * sparse_value, axis=1) * SCALE
-        probabilities = tl.where(valid, tl.exp(scores - lse), 0.0)
-        grad_probs = tl.sum(grad_output * sparse_value, axis=1)
-        grad_scores = probabilities * (grad_probs - delta)
-        grad_query += grad_scores[:, None] * sparse_value * SCALE
-
-    first_key = query_block * BLOCK_M - WINDOW + 1
-    offsets_n_base = tl.arange(0, BLOCK_N)
-    for key_tile in tl.range(0, NUM_LOCAL_TILES):
-        local_start = first_key + key_tile * BLOCK_N
-        offsets_n = local_start + offsets_n_base
-        key_mask = (offsets_n >= 0) & (offsets_n < S)
-        local_values = _load_bhsd(
-            local_kv_ptr,
-            LOCAL_KV_STRIDES,
-            batch,
-            head,
-            offsets_n,
-            offsets_d,
-            key_mask[:, None] & dimension_mask[None, :],
-        )
-        scores = tl.dot(query, tl.trans(local_values), input_precision="tf32x3") * SCALE
-        valid = (
-            query_mask[:, None]
-            & key_mask[None, :]
-            & (offsets_n[None, :] <= offsets_m[:, None])
-            & (offsets_n[None, :] >= offsets_m[:, None] - WINDOW + 1)
-        )
-        if HAS_DOC_IDS:
-            key_doc_ids = _load_bs(doc_ids_ptr, DOC_IDS_STRIDES, batch, offsets_n, key_mask, -2)
-            doc_mask = query_doc_ids[:, None] == key_doc_ids[None, :]
-            valid = valid & doc_mask
-
-        probabilities = tl.exp(scores - lse[:, None])
-        probabilities = tl.where(valid, probabilities, 0.0)
-        grad_probabilities = tl.dot(grad_output, tl.trans(local_values), input_precision="tf32x3")
-        grad_scores = probabilities * (grad_probabilities - delta[:, None])
-        grad_query += (
-            tl.dot(
-                grad_scores.to(local_values.dtype),
-                local_values,
-                input_precision="tf32x3",
-            )
-            * SCALE
-        )
-
-    _store_bhsd(
-        grad_query_ptr,
-        grad_query,
-        QUERY_STRIDES,
-        batch,
-        head,
-        offsets_m,
-        offsets_d,
-        matrix_mask,
-    )
-
-    # Sink gradient
-    sink = tl.load(attention_sink_ptr + head)
-    sink_probability = tl.exp(sink - lse)
-    sink_gradient = tl.where(query_mask, -sink_probability * delta, 0.0)
-    tl.atomic_add(grad_sink_ptr + head, tl.sum(sink_gradient, axis=0))
-
-
-@triton.jit
-def _selected_attention_bwd_dlocal_kv(
-    query_ptr,
-    local_kv_ptr,
-    doc_ids_ptr,
-    output_ptr,
-    grad_output_ptr,
-    lse_ptr,
-    grad_local_kv_ptr,
-    QUERY_STRIDES: tl.constexpr,
-    LOCAL_KV_STRIDES: tl.constexpr,
-    GRAD_LOCAL_KV_STRIDES: tl.constexpr,
-    DOC_IDS_STRIDES: tl.constexpr,
-    LSE_STRIDES: tl.constexpr,
-    H: tl.constexpr,
-    S: tl.constexpr,
-    D: tl.constexpr,
-    WINDOW: tl.constexpr,
-    SCALE: tl.constexpr,
-    HAS_DOC_IDS: tl.constexpr,
-    NUM_QUERY_TILES: tl.constexpr,
-    BLOCK_M: tl.constexpr,
-    BLOCK_N: tl.constexpr,
-    BLOCK_D: tl.constexpr,
-):
-    """Backward for dKV (local sliding window)."""
-    key_block = tl.program_id(0)
-    batch_head = tl.program_id(1)
-    head = batch_head % H
-    batch = batch_head // H
-
-    offsets_n = key_block * BLOCK_N + tl.arange(0, BLOCK_N)
-    offsets_d = tl.arange(0, BLOCK_D)
-    key_mask = offsets_n < S
-    dimension_mask = offsets_d < D
-
-    local_values = _load_bhsd(
-        local_kv_ptr,
-        LOCAL_KV_STRIDES,
-        batch,
-        head,
-        offsets_n,
-        offsets_d,
-        key_mask[:, None] & dimension_mask[None, :],
-    )
-
-    if HAS_DOC_IDS:
-        key_doc_ids = _load_bs(doc_ids_ptr, DOC_IDS_STRIDES, batch, offsets_n, key_mask, -2)
-
-    grad_values = tl.zeros((BLOCK_N, BLOCK_D), tl.float32)
-    first_query = key_block * BLOCK_N
-    offsets_m_base = tl.arange(0, BLOCK_M)
-
-    for query_tile in tl.range(0, NUM_QUERY_TILES):
-        query_start = first_query + query_tile * BLOCK_M
-        offsets_m = query_start + offsets_m_base
-        query_mask = offsets_m < S
-        matrix_mask = query_mask[:, None] & dimension_mask[None, :]
-        query = _load_bhsd(
-            query_ptr, QUERY_STRIDES, batch, head, offsets_m, offsets_d, matrix_mask
-        )
-        output = _load_bhsd(
-            output_ptr, QUERY_STRIDES, batch, head, offsets_m, offsets_d, matrix_mask
-        )
-        grad_output = _load_bhsd(
-            grad_output_ptr, QUERY_STRIDES, batch, head, offsets_m, offsets_d, matrix_mask
-        )
-        lse = tl.load(
-            lse_ptr + batch * LSE_STRIDES[0] + head * LSE_STRIDES[1] + offsets_m * LSE_STRIDES[2],
-            mask=query_mask,
-            other=0.0,
-        )
-        delta = tl.sum(grad_output * output, axis=1)
-        valid = (
-            query_mask[:, None]
-            & key_mask[None, :]
-            & (offsets_n[None, :] <= offsets_m[:, None])
-            & (offsets_n[None, :] >= offsets_m[:, None] - WINDOW + 1)
-        )
-        if HAS_DOC_IDS:
-            query_doc_ids = _load_bs(
-                doc_ids_ptr, DOC_IDS_STRIDES, batch, offsets_m, query_mask, -1
-            )
-            doc_mask = query_doc_ids[:, None] == key_doc_ids[None, :]
-            valid = valid & doc_mask
-
-        scores = tl.dot(query, tl.trans(local_values), input_precision="tf32x3") * SCALE
-        probabilities = tl.exp(scores - lse[:, None])
-        probabilities = tl.where(valid, probabilities, 0.0)
-        grad_probabilities = tl.dot(grad_output, tl.trans(local_values), input_precision="tf32x3")
-        grad_scores = probabilities * (grad_probabilities - delta[:, None])
-        grad_values += tl.dot(
-            tl.trans(probabilities.to(grad_output.dtype)),
-            grad_output,
-            input_precision="tf32x3",
-        )
-        grad_values += (
-            tl.dot(
-                tl.trans(grad_scores.to(query.dtype)),
-                query,
-                input_precision="tf32x3",
-            )
-            * SCALE
-        )
-
-    _store_bhsd(
-        grad_local_kv_ptr,
-        grad_values,
-        GRAD_LOCAL_KV_STRIDES,
-        batch,
-        head,
-        offsets_n,
-        offsets_d,
-        key_mask[:, None] & dimension_mask[None, :],
-    )
-
-
-@triton.autotune(
-    configs=[
-        triton.Config({}, num_warps=num_warps, num_stages=num_stages)
-        for num_warps in (4, 8)
-        for num_stages in (1, 3)
-    ],
-    key=["H", "S", "D", "WINDOW", "HAS_DOC_IDS"],
-    cache_results=True,
-)
-@triton.jit
-def _selected_attention_bwd_dlocal_kv_tma(
-    query_desc,
-    local_desc,
-    doc_ids_ptr,
-    output_desc,
-    grad_output_desc,
-    lse_ptr,
-    grad_local_desc,
-    DOC_IDS_STRIDES: tl.constexpr,
-    LSE_STRIDES: tl.constexpr,
-    H: tl.constexpr,
-    S: tl.constexpr,
-    D: tl.constexpr,
-    WINDOW: tl.constexpr,
-    SCALE: tl.constexpr,
-    HAS_DOC_IDS: tl.constexpr,
-    NUM_QUERY_TILES: tl.constexpr,
-    BLOCK_M: tl.constexpr,
-    BLOCK_N: tl.constexpr,
-    BLOCK_D: tl.constexpr,
-):
-    """TMA backward for local-KV gradients."""
-    key_block = tl.program_id(0)
-    batch_head = tl.program_id(1)
-    head = batch_head % H
-    batch = batch_head // H
-
-    offsets_n = key_block * BLOCK_N + tl.arange(0, BLOCK_N)
-    key_mask = offsets_n < S
-    local_values = tl.reshape(
-        local_desc.load([batch, head, key_block * BLOCK_N, 0]),
-        (BLOCK_N, BLOCK_D),
-    )
-
-    if HAS_DOC_IDS:
-        key_doc_ids = _load_bs(doc_ids_ptr, DOC_IDS_STRIDES, batch, offsets_n, key_mask, -2)
-
-    grad_values = tl.zeros((BLOCK_N, BLOCK_D), tl.float32)
-    first_query = key_block * BLOCK_N
-    offsets_m_base = tl.arange(0, BLOCK_M)
-
-    for query_tile in tl.range(0, NUM_QUERY_TILES):
-        query_start = first_query + query_tile * BLOCK_M
-        offsets_m = query_start + offsets_m_base
-        query_mask = offsets_m < S
-        query = tl.reshape(
-            query_desc.load([batch, head, query_start, 0]),
-            (BLOCK_M, BLOCK_D),
-        )
-        output = tl.reshape(
-            output_desc.load([batch, head, query_start, 0]),
-            (BLOCK_M, BLOCK_D),
-        )
-        grad_output = tl.reshape(
-            grad_output_desc.load([batch, head, query_start, 0]),
-            (BLOCK_M, BLOCK_D),
-        )
-        lse = tl.load(
-            lse_ptr + batch * LSE_STRIDES[0] + head * LSE_STRIDES[1] + offsets_m * LSE_STRIDES[2],
-            mask=query_mask,
-            other=0.0,
-        )
-        delta = tl.sum(grad_output * output, axis=1)
-        valid = (
-            query_mask[:, None]
-            & key_mask[None, :]
-            & (offsets_n[None, :] <= offsets_m[:, None])
-            & (offsets_n[None, :] >= offsets_m[:, None] - WINDOW + 1)
-        )
-        if HAS_DOC_IDS:
-            query_doc_ids = _load_bs(
-                doc_ids_ptr, DOC_IDS_STRIDES, batch, offsets_m, query_mask, -1
-            )
-            valid &= query_doc_ids[:, None] == key_doc_ids[None, :]
-
-        scores = tl.dot(query, tl.trans(local_values), input_precision="tf32x3") * SCALE
-        probabilities = tl.where(valid, tl.exp(scores - lse[:, None]), 0.0)
-        grad_probabilities = tl.dot(grad_output, tl.trans(local_values), input_precision="tf32x3")
-        grad_scores = probabilities * (grad_probabilities - delta[:, None])
-        grad_values += tl.dot(
-            tl.trans(probabilities.to(grad_output.dtype)),
-            grad_output,
-            input_precision="tf32x3",
-        )
-        grad_values += (
-            tl.dot(
-                tl.trans(grad_scores.to(query.dtype)),
-                query,
-                input_precision="tf32x3",
-            )
-            * SCALE
-        )
-
-    grad_local_desc.store(
-        [batch, head, key_block * BLOCK_N, 0],
-        tl.reshape(grad_values, (1, 1, BLOCK_N, BLOCK_D)),
-    )
-
-
-@triton.autotune(
-    configs=[
-        triton.Config(
-            {"BLOCK_M": block_m},
-            num_warps=num_warps,
-            num_stages=num_stages,
-        )
-        for block_m in (16, 32, 64)
-        for num_warps in (4, 8)
-        for num_stages in (1, 3)
-    ],
-    key=["H", "S", "D", "SPARSE_SEQ_LEN", "TOPK"],
-    cache_results=True,
-)
-@triton.jit
-def _selected_attention_bwd_dsparse_kv(
-    query_ptr,
-    sparse_kv_ptr,
-    selected_queries_ptr,
-    block_offsets_ptr,
-    output_ptr,
-    grad_output_ptr,
-    lse_ptr,
-    grad_sparse_kv_ptr,
-    QUERY_STRIDES: tl.constexpr,
-    SPARSE_KV_STRIDES: tl.constexpr,
-    SELECTED_QUERIES_STRIDES: tl.constexpr,
-    BLOCK_OFFSETS_STRIDES: tl.constexpr,
-    LSE_STRIDES: tl.constexpr,
-    GRAD_SPARSE_KV_STRIDES: tl.constexpr,
-    H: tl.constexpr,
-    S: tl.constexpr,
-    D: tl.constexpr,
-    SPARSE_SEQ_LEN: tl.constexpr,
-    TOPK: tl.constexpr,
-    SCALE: tl.constexpr,
-    BLOCK_M: tl.constexpr,
-    BLOCK_D: tl.constexpr,
-):
-    """Compute sparse-KV gradients from an inverted index of selecting queries."""
-    sparse_index = tl.program_id(0)
-    batch_head = tl.program_id(1)
-    head = batch_head % H
-    batch = batch_head // H
-
-    offsets_m_base = tl.arange(0, BLOCK_M)
-    offsets_d = tl.arange(0, BLOCK_D)
-    dot_rows = tl.arange(0, 16)
-    dimension_mask = offsets_d < D
-
-    sparse_value = tl.load(
-        sparse_kv_ptr
-        + batch * SPARSE_KV_STRIDES[0]
-        + head * SPARSE_KV_STRIDES[1]
-        + sparse_index * SPARSE_KV_STRIDES[2]
-        + offsets_d * SPARSE_KV_STRIDES[3],
-        mask=dimension_mask,
-        other=0.0,
-    )
-    dot_value = tl.where(
-        dot_rows[None, :] == 0,
-        sparse_value[:, None],
-        0.0,
-    )
-    grad_value = tl.zeros((16, BLOCK_D), tl.float32)
-    entry_start = tl.load(
-        block_offsets_ptr
-        + batch * BLOCK_OFFSETS_STRIDES[0]
-        + sparse_index * BLOCK_OFFSETS_STRIDES[1]
-    )
-    entry_end = tl.load(
-        block_offsets_ptr
-        + batch * BLOCK_OFFSETS_STRIDES[0]
-        + (sparse_index + 1) * BLOCK_OFFSETS_STRIDES[1]
-    )
-
-    for entry_tile in tl.range(entry_start, entry_end, BLOCK_M):
-        entry_offsets = entry_tile + offsets_m_base
-        entry_mask = entry_offsets < entry_end
-        query_positions = tl.load(
-            selected_queries_ptr
-            + batch * SELECTED_QUERIES_STRIDES[0]
-            + entry_offsets * SELECTED_QUERIES_STRIDES[1],
-            mask=entry_mask,
-            other=0,
-        )
-        query_mask = entry_mask & (query_positions >= 0) & (query_positions < S)
-        matrix_mask = query_mask[:, None] & dimension_mask[None, :]
-        query = _load_bhsd(
-            query_ptr, QUERY_STRIDES, batch, head, query_positions, offsets_d, matrix_mask
-        )
-        output = _load_bhsd(
-            output_ptr, QUERY_STRIDES, batch, head, query_positions, offsets_d, matrix_mask
-        )
-        grad_output = _load_bhsd(
-            grad_output_ptr, QUERY_STRIDES, batch, head, query_positions, offsets_d, matrix_mask
-        )
-        lse = tl.load(
-            lse_ptr
-            + batch * LSE_STRIDES[0]
-            + head * LSE_STRIDES[1]
-            + query_positions * LSE_STRIDES[2],
-            mask=query_mask,
-            other=0.0,
-        )
-        score_tile = tl.dot(query, dot_value, input_precision="tf32x3")
-        scores = tl.sum(score_tile * (dot_rows[None, :] == 0), axis=1) * SCALE
-        grad_probability_tile = tl.dot(grad_output, dot_value, input_precision="tf32x3")
-        grad_probabilities = tl.sum(grad_probability_tile * (dot_rows[None, :] == 0), axis=1)
-        delta = tl.sum(grad_output * output, axis=1)
-        probabilities = tl.where(query_mask, tl.exp(scores - lse), 0.0)
-        grad_scores = probabilities * (grad_probabilities - delta)
-        combined_weights = tl.cat(probabilities, grad_scores * SCALE, dim=0)
-        combined_values = tl.cat(grad_output, query, dim=0)
-        combined_weight_tile = tl.where(
-            dot_rows[:, None] == 0,
-            combined_weights[None, :],
-            0.0,
-        )
-        grad_value = tl.dot(
-            combined_weight_tile.to(query.dtype),
-            combined_values,
-            acc=grad_value,
-            input_precision="tf32x3",
-        )
-
-    tl.store(
-        grad_sparse_kv_ptr
-        + batch * GRAD_SPARSE_KV_STRIDES[0]
-        + head * GRAD_SPARSE_KV_STRIDES[1]
-        + sparse_index * GRAD_SPARSE_KV_STRIDES[2]
-        + offsets_d * GRAD_SPARSE_KV_STRIDES[3],
-        tl.sum(grad_value * (dot_rows[:, None] == 0), axis=0),
-        mask=dimension_mask,
-    )
-
-
-def _launch_forward(
-    query: torch.Tensor,
-    sparse_kv: torch.Tensor,
-    local_kv: torch.Tensor,
-    kv_indices: torch.Tensor,
-    attention_sink: torch.Tensor,
-    doc_ids: torch.Tensor | None,
-    sliding_window_size: int,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Launch the forward kernel and return its output and log-sum-exp state."""
-    batch, heads, seq_len, head_dim = query.shape
-    topk = kv_indices.shape[-1]
-    sparse_seq_len = sparse_kv.shape[2]
-    block_d = max(16, triton.next_power_of_2(head_dim))
-    output = torch.empty_like(query)
-    lse = torch.empty(batch, heads, seq_len, device=query.device, dtype=torch.float32)
-    has_doc_ids = doc_ids is not None
-    doc_ids = query if doc_ids is None else doc_ids
-
-    # This head-major schedule is tuned for shared KV on Blackwell.
-    if (
-        torch.cuda.get_device_capability(query.device)[0] >= 10
-        and query.dtype == torch.bfloat16
-        and sparse_kv.stride(1) == 0
-        and local_kv.stride(1) == 0
-        and query.stride(-1) == 1
-        and sparse_kv.stride(-1) == 1
-        and local_kv.stride(-1) == 1
-        and 16 <= heads <= 128
-        and head_dim % 16 == 0
-        and head_dim <= 128
-        and topk <= 64
-        and sliding_window_size <= 2048
-    ):
-        block_h = max(16, triton.next_power_of_2(heads))
-        block_k = max(16, triton.next_power_of_2(topk)) if topk else 16
-        _selected_attention_fwd_shared[(seq_len, batch, triton.cdiv(heads, block_h))](
-            query,
-            sparse_kv,
-            local_kv,
-            kv_indices,
-            doc_ids,
-            attention_sink,
-            output,
-            lse,
-            QUERY_STRIDES=query.stride(),
-            SPARSE_KV_STRIDES=sparse_kv.stride(),
-            LOCAL_KV_STRIDES=local_kv.stride(),
-            KV_INDICES_STRIDES=kv_indices.stride(),
-            DOC_IDS_STRIDES=doc_ids.stride(),
-            LSE_STRIDES=lse.stride(),
-            B=batch,
-            H=heads,
-            S=seq_len,
-            D=head_dim,
-            SPARSE_SEQ_LEN=sparse_seq_len,
-            TOPK=topk,
-            WINDOW=sliding_window_size,
-            SCALE=1.0 / math.sqrt(head_dim),
-            HAS_DOC_IDS=has_doc_ids,
-            BLOCK_H=block_h,
-            BLOCK_K=block_k,
-            BLOCK_D=block_d,
-        )
-        return output, lse
-
-    block_m = 64
-    block_n = 128
-    num_local_tiles = (
-        triton.cdiv(sliding_window_size + block_m - 1, block_n) if sliding_window_size else 0
-    )
-    grid = (triton.cdiv(seq_len, block_m), batch * heads)
-    if can_use_tma(query) and can_use_tma(local_kv):
-        query_desc = TensorDescriptor.from_tensor(query, [1, 1, block_m, block_d])
-        local_desc = TensorDescriptor.from_tensor(local_kv, [1, 1, block_n, block_d])
-        output_desc = TensorDescriptor.from_tensor(output, [1, 1, block_m, block_d])
-        _selected_attention_fwd_tma[grid](
-            query_desc,
-            sparse_kv,
-            local_desc,
-            kv_indices,
-            doc_ids,
-            attention_sink,
-            output_desc,
-            lse,
-            SPARSE_KV_STRIDES=sparse_kv.stride(),
-            KV_INDICES_STRIDES=kv_indices.stride(),
-            DOC_IDS_STRIDES=doc_ids.stride(),
-            LSE_STRIDES=lse.stride(),
-            H=heads,
-            S=seq_len,
-            D=head_dim,
-            SPARSE_SEQ_LEN=sparse_seq_len,
-            TOPK=topk,
-            WINDOW=sliding_window_size,
-            SCALE=1.0 / math.sqrt(head_dim),
-            HAS_DOC_IDS=has_doc_ids,
-            NUM_LOCAL_TILES=num_local_tiles,
-            BLOCK_M=block_m,
-            BLOCK_N=block_n,
-            BLOCK_D=block_d,
-        )
-    else:
-        _selected_attention_fwd[grid](
-            query,
-            sparse_kv,
-            local_kv,
-            kv_indices,
-            doc_ids,
-            attention_sink,
-            output,
-            lse,
-            QUERY_STRIDES=query.stride(),
-            SPARSE_KV_STRIDES=sparse_kv.stride(),
-            LOCAL_KV_STRIDES=local_kv.stride(),
-            KV_INDICES_STRIDES=kv_indices.stride(),
-            DOC_IDS_STRIDES=doc_ids.stride(),
-            LSE_STRIDES=lse.stride(),
-            H=heads,
-            S=seq_len,
-            D=head_dim,
-            SPARSE_SEQ_LEN=sparse_seq_len,
-            TOPK=topk,
-            WINDOW=sliding_window_size,
-            SCALE=1.0 / math.sqrt(head_dim),
-            HAS_DOC_IDS=has_doc_ids,
-            NUM_LOCAL_TILES=num_local_tiles,
-            BLOCK_M=block_m,
-            BLOCK_N=block_n,
-            BLOCK_D=block_d,
-            num_warps=8,
-            num_stages=3,
-        )
-    return output, lse
-
-
-def _build_index_query_map(
-    kv_indices: torch.Tensor,
-    sparse_seq_len: int,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Invert query-major indices into block-major query lists (CSR format)."""
-    batch, _seq_len, topk = kv_indices.shape
-    if topk == 0:
-        selected_queries = torch.empty(batch, 0, device=kv_indices.device, dtype=torch.int32)
-        block_offsets = torch.zeros(
-            batch, sparse_seq_len + 1, device=kv_indices.device, dtype=torch.int32
-        )
-        return selected_queries, block_offsets
-
-    sorted_indices, sorted_entries = torch.sort(kv_indices.flatten(1), dim=-1)
-    selected_queries = torch.div(sorted_entries, topk, rounding_mode="floor").to(torch.int32)
-    block_ids = torch.arange(
-        sparse_seq_len + 1, device=kv_indices.device, dtype=sorted_indices.dtype
-    )
-    block_ids = block_ids.unsqueeze(0).expand(batch, -1).contiguous()
-    block_offsets = torch.searchsorted(sorted_indices, block_ids).to(torch.int32)
-    return selected_queries.contiguous(), block_offsets.contiguous()
-
-
-def _launch_backward(
-    query: torch.Tensor,
-    sparse_kv: torch.Tensor,
-    local_kv: torch.Tensor,
-    kv_indices: torch.Tensor,
-    selected_queries: torch.Tensor,
-    block_offsets: torch.Tensor,
-    attention_sink: torch.Tensor,
-    doc_ids: torch.Tensor | None,
-    output: torch.Tensor,
-    lse: torch.Tensor,
-    grad_output: torch.Tensor,
-    sliding_window_size: int,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
-    """Launch backward kernels for selected attention."""
-    batch, heads, seq_len, head_dim = query.shape
-    sparse_seq_len = sparse_kv.shape[2]
-    topk = kv_indices.shape[-1]
-    block_m = 64
-    block_n = 32
-    block_d = max(16, triton.next_power_of_2(head_dim))
-    scale = 1.0 / math.sqrt(head_dim)
-    grad_output = grad_output.contiguous()
-
-    grad_query = torch.empty_like(query)
-    # Expanded shared KV has a zero head stride, but its per-head gradients
-    # need distinct storage before ExpandBackward sums them.
-    grad_local_kv = torch.empty(local_kv.shape, device=local_kv.device, dtype=local_kv.dtype)
-    grad_sink_fp32 = torch.zeros(heads, device=query.device, dtype=torch.float32)
-
-    has_doc_ids = doc_ids is not None
-    doc_ids = query if doc_ids is None else doc_ids
-    use_tma = can_use_tma(query) and can_use_tma(local_kv) and can_use_tma(grad_output)
-
-    num_local_key_tiles = (
-        triton.cdiv(sliding_window_size + block_m - 1, block_n) if sliding_window_size else 0
-    )
-    num_local_query_tiles = (
-        triton.cdiv(sliding_window_size + block_n - 1, block_m) if sliding_window_size else 0
-    )
-
-    _selected_attention_bwd_dq[(triton.cdiv(seq_len, block_m), batch * heads)](
-        query,
-        sparse_kv,
-        local_kv,
-        kv_indices,
-        doc_ids,
-        output,
-        grad_output,
-        lse,
-        attention_sink,
-        grad_query,
-        grad_sink_fp32,
-        QUERY_STRIDES=query.stride(),
-        SPARSE_KV_STRIDES=sparse_kv.stride(),
-        LOCAL_KV_STRIDES=local_kv.stride(),
-        KV_INDICES_STRIDES=kv_indices.stride(),
-        DOC_IDS_STRIDES=doc_ids.stride(),
-        LSE_STRIDES=lse.stride(),
-        H=heads,
-        S=seq_len,
-        D=head_dim,
-        SPARSE_SEQ_LEN=sparse_seq_len,
-        TOPK=topk,
-        WINDOW=sliding_window_size,
-        SCALE=scale,
-        HAS_DOC_IDS=has_doc_ids,
-        NUM_LOCAL_TILES=num_local_key_tiles,
-        BLOCK_M=block_m,
-        BLOCK_N=block_n,
-        BLOCK_D=block_d,
-        num_warps=8,
-        num_stages=3,
-    )
-
-    local_grid = (triton.cdiv(seq_len, block_n), batch * heads)
-    if use_tma:
-        query_desc = TensorDescriptor.from_tensor(query, [1, 1, block_m, block_d])
-        local_desc = TensorDescriptor.from_tensor(local_kv, [1, 1, block_n, block_d])
-        output_desc = TensorDescriptor.from_tensor(output, [1, 1, block_m, block_d])
-        grad_output_desc = TensorDescriptor.from_tensor(grad_output, [1, 1, block_m, block_d])
-        grad_local_desc = TensorDescriptor.from_tensor(grad_local_kv, [1, 1, block_n, block_d])
-        _selected_attention_bwd_dlocal_kv_tma[local_grid](
-            query_desc,
-            local_desc,
-            doc_ids,
-            output_desc,
-            grad_output_desc,
-            lse,
-            grad_local_desc,
-            DOC_IDS_STRIDES=doc_ids.stride(),
-            LSE_STRIDES=lse.stride(),
-            H=heads,
-            S=seq_len,
-            D=head_dim,
-            WINDOW=sliding_window_size,
-            SCALE=scale,
-            HAS_DOC_IDS=has_doc_ids,
-            NUM_QUERY_TILES=num_local_query_tiles,
-            BLOCK_M=block_m,
-            BLOCK_N=block_n,
-            BLOCK_D=block_d,
-        )
-    else:
-        _selected_attention_bwd_dlocal_kv[local_grid](
-            query,
-            local_kv,
-            doc_ids,
-            output,
-            grad_output,
-            lse,
-            grad_local_kv,
-            QUERY_STRIDES=query.stride(),
-            LOCAL_KV_STRIDES=local_kv.stride(),
-            GRAD_LOCAL_KV_STRIDES=grad_local_kv.stride(),
-            DOC_IDS_STRIDES=doc_ids.stride(),
-            LSE_STRIDES=lse.stride(),
-            H=heads,
-            S=seq_len,
-            D=head_dim,
-            WINDOW=sliding_window_size,
-            SCALE=scale,
-            HAS_DOC_IDS=has_doc_ids,
-            NUM_QUERY_TILES=num_local_query_tiles,
-            BLOCK_M=block_m,
-            BLOCK_N=block_n,
-            BLOCK_D=block_d,
-            num_warps=8,
-            num_stages=3,
-        )
-
-    if topk > 0:
-        grad_sparse_kv = torch.empty(
-            sparse_kv.shape, device=sparse_kv.device, dtype=sparse_kv.dtype
-        )
-        _selected_attention_bwd_dsparse_kv[(sparse_seq_len, batch * heads)](
-            query,
-            sparse_kv,
-            selected_queries,
-            block_offsets,
-            output,
-            grad_output,
-            lse,
-            grad_sparse_kv,
-            QUERY_STRIDES=query.stride(),
-            SPARSE_KV_STRIDES=sparse_kv.stride(),
-            SELECTED_QUERIES_STRIDES=selected_queries.stride(),
-            BLOCK_OFFSETS_STRIDES=block_offsets.stride(),
-            LSE_STRIDES=lse.stride(),
-            GRAD_SPARSE_KV_STRIDES=grad_sparse_kv.stride(),
-            H=heads,
-            S=seq_len,
-            D=head_dim,
-            SPARSE_SEQ_LEN=sparse_seq_len,
-            TOPK=topk,
-            SCALE=scale,
-            BLOCK_D=block_d,
-        )
-    else:
-        grad_sparse_kv = torch.zeros(
-            sparse_kv.shape, device=sparse_kv.device, dtype=sparse_kv.dtype
-        )
-
-    return grad_query, grad_sparse_kv, grad_local_kv, grad_sink_fp32.to(attention_sink.dtype)
-
-
-class _SelectedAttentionFunction(torch.autograd.Function):
-    """Autograd wrapper around the Triton launchers."""
-
-    @staticmethod
-    def forward(
-        ctx,
-        query: torch.Tensor,
-        sparse_kv: torch.Tensor,
-        local_kv: torch.Tensor,
-        kv_indices: torch.Tensor,
-        attention_sink: torch.Tensor,
-        doc_ids: torch.Tensor | None,
-        sliding_window_size: int,
-    ) -> torch.Tensor:
-        output, lse = _launch_forward(
-            query,
-            sparse_kv,
-            local_kv,
-            kv_indices,
-            attention_sink,
-            doc_ids,
-            sliding_window_size,
-        )
-        selected_queries, block_offsets = _build_index_query_map(kv_indices, sparse_kv.shape[2])
-        ctx.save_for_backward(
-            query,
-            sparse_kv,
-            local_kv,
-            kv_indices,
-            selected_queries,
-            block_offsets,
-            attention_sink,
-            output,
-            lse,
-        )
-        ctx.doc_ids = doc_ids
-        ctx.sliding_window_size = sliding_window_size
-        return output
-
-    @staticmethod
-    def backward(ctx, grad_output: torch.Tensor):
-        (
-            query,
-            sparse_kv,
-            local_kv,
-            kv_indices,
-            selected_queries,
-            block_offsets,
-            attention_sink,
-            output,
-            lse,
-        ) = ctx.saved_tensors
-        grad_query, grad_sparse_kv, grad_local_kv, grad_sink = _launch_backward(
-            query,
-            sparse_kv,
-            local_kv,
-            kv_indices,
-            selected_queries,
-            block_offsets,
-            attention_sink,
-            ctx.doc_ids,
-            output,
-            lse,
-            grad_output,
-            ctx.sliding_window_size,
-        )
-        return grad_query, grad_sparse_kv, grad_local_kv, None, grad_sink, None, None
-
-
-def selected_attention(
-    query: torch.Tensor,
-    local_kv: torch.Tensor,
-    sparse_kv: torch.Tensor,
-    kv_indices: torch.Tensor,
-    attention_sink: torch.Tensor,
-    doc_ids: torch.Tensor | None,
-    sliding_window_size: int,
-    share_kv: bool = True,
-) -> torch.Tensor:
-    """Triton implementation of selected attention.
-
-    Args:
-        query: (batch, heads, seq_len, head_dim) — queries.
-        local_kv: (batch, 1 or heads, seq_len, head_dim) — local sliding-window key-values.
-        sparse_kv: (batch, 1 or heads, sparse_seq_len, head_dim) — candidate KV pool.
-        kv_indices: (batch, seq_len, topk) — which sparse_kv positions each query attends to.
-        attention_sink: (heads,) — learned per-head sink weight.
-        doc_ids: (batch, seq_len) or None — document IDs for packing isolation.
-        sliding_window_size: size of the causal sliding window.
-        share_kv: if True, expand single-head KV to all heads.
-
-    Returns:
-        Attention output with same shape as query.
-    """
-    heads = query.shape[1]
-
-    if query.device.type != "cuda":
-        raise ValueError("The Triton selected attention backend requires CUDA tensors.")
-
-    # Expand shared KV heads (stride-zero broadcast; no memory duplication)
-    if share_kv:
-        local_kv = local_kv.expand(-1, heads, -1, -1)
-        sparse_kv = sparse_kv.expand(-1, heads, -1, -1)
-
-    query = query.contiguous()
-    kv_indices = kv_indices.contiguous()
-    if doc_ids is not None:
-        doc_ids = doc_ids.contiguous()
-
-    if torch.is_grad_enabled() and any(
-        t.requires_grad for t in (query, local_kv, sparse_kv, attention_sink)
-    ):
-        return _SelectedAttentionFunction.apply(
-            query, sparse_kv, local_kv, kv_indices, attention_sink, doc_ids, sliding_window_size
-        )
-    return _launch_forward(
-        query, sparse_kv, local_kv, kv_indices, attention_sink, doc_ids, sliding_window_size
-    )[0]
